Change quotes formatting rules

+++ eslint.config.js
@@ -23,7 +23,7 @@ export default [
     rules: {
       indent: ["error", 2, { SwitchCase: 1 }],
       "linebreak-style": ["error", "unix"],
-      quotes: ["error", "single"],
+      quotes: ["error", "double"],
       semi: ["error", "always"],
       "no-unused-vars": ["error", { argsIgnorePattern: "^_" }],
       "no-console": "off",
